fix cmd
